+++ frontend/src/components/dependency-graph/types.ts
@@ -1,3 +1,17 @@
 /* Copyright 2024 Marimo. All rights reserved. */
+import { CellId } from "@/core/cells/ids";
+
 export type LayoutDirection = "TB" | "LR";
 export type GraphLayoutView = LayoutDirection | "_minimap_";
+
+export type GraphSelection =
+  | {
+      type: "node";
+      id: CellId;
+    }
+  | {
+      type: "edge";
+      source: CellId;
+      target: CellId;
+    }
+  | undefined;
